rsync by default #115 + fix #119

+++ uninstall.sh
@@ -1,7 +1,13 @@
 #!/usr/bin/env sh
 
+if dpkg -l log2ram ; then
+  echo "Please run : apt remove log2ram"
+  exit 1
+fi
+
 if [ "$(id -u)" -eq 0 ]
 then
+  echo "Not apt installed. Remove will continue with this script..."
   service log2ram stop
   systemctl disable log2ram
   rm /etc/systemd/system/log2ram.service
